refactor: reorganize service packages
